+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support bounds for main trait parameters
 - Support const parameters in main trait
+- Keep original trait param identifiers
 
 ## [0.5.0] - 2023-11-23